backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.24.0
openai>=1.0.0
pdfplumber>=0.10.0
python-multipart>=0.0.6
pydantic>=2.0.0
python-dotenv>=1.0.0
